requirements.txt
boto3==1.34.131
fastapi==0.111.1
mysql-connector-python==8.4.0
python-dotenv==1.0.1
uvicorn==0.30.3
